requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.109.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
priority==2.0.0
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.36.2
tenacity==8.2.3
typing_extensions==4.9.0
urllib3==2.1.0
wsproto==1.2.0
google-generativeai==0.3.1
